add ConfigFile abstract class
