--- src/fspm/models/config/OrganConfig.java
@@ -1,4 +1,0 @@
-package fspm.models.config;
-
-public class OrganConfig {
-}
