+++ src/fspm/models/config/OrganConfigFile.java
@@ -1,0 +1,4 @@
+package fspm.models.config;
+
+public class OrganConfigFile extends ConfigFile {
+}
