+++ src/fspm/utils/Utils.rgg
@@ -1,3 +1,5 @@
+package fspm.utils;
+
 // Temporary class to allow printing regardless of file type (rgg, java or xl)
 
 public static void print(String s) {
